Adds helper access to underlying modules
