+++ node/src/lib.rs
@@ -20,6 +20,14 @@ extern crate async_trait;
 #[macro_use]
 extern crate tracing;
 
+pub use snarkos_node_cdn as cdn;
+pub use snarkos_node_consensus as consensus;
+pub use snarkos_node_messages as messages;
+pub use snarkos_node_rest as rest;
+pub use snarkos_node_router as router;
+pub use snarkos_node_tcp as tcp;
+pub use snarkvm;
+
 mod beacon;
 pub use beacon::*;
 
@@ -32,131 +40,11 @@ pub use prover::*;
 mod validator;
 pub use validator::*;
 
+mod node;
+pub use node::*;
+
 mod helpers;
+pub use helpers::*;
 
 mod traits;
 pub use traits::*;
-
-pub use snarkos_node_messages::NodeType;
-
-use snarkos_account::Account;
-use snarkvm::{
-    prelude::{Address, Block, Network, PrivateKey, ViewKey},
-    synthesizer::store::helpers::{memory::ConsensusMemory, rocksdb::ConsensusDB},
-};
-
-use anyhow::Result;
-use std::{net::SocketAddr, sync::Arc};
-
-pub enum Node<N: Network> {
-    /// A beacon is a full node, capable of producing blocks.
-    Beacon(Arc<Beacon<N, ConsensusDB<N>>>),
-    /// A validator is a full node, capable of validating blocks.
-    Validator(Arc<Validator<N, ConsensusDB<N>>>),
-    /// A prover is a full node, capable of producing proofs for consensus.
-    Prover(Arc<Prover<N, ConsensusMemory<N>>>),
-    /// A client node is a full node, capable of querying with the network.
-    Client(Arc<Client<N, ConsensusMemory<N>>>),
-}
-
-impl<N: Network> Node<N> {
-    /// Initializes a new beacon node.
-    pub async fn new_beacon(
-        node_ip: SocketAddr,
-        rest_ip: Option<SocketAddr>,
-        account: Account<N>,
-        trusted_peers: &[SocketAddr],
-        genesis: Block<N>,
-        cdn: Option<String>,
-        dev: Option<u16>,
-    ) -> Result<Self> {
-        Ok(Self::Beacon(Arc::new(Beacon::new(node_ip, rest_ip, account, trusted_peers, genesis, cdn, dev).await?)))
-    }
-
-    /// Initializes a new validator node.
-    pub async fn new_validator(
-        node_ip: SocketAddr,
-        rest_ip: Option<SocketAddr>,
-        account: Account<N>,
-        trusted_peers: &[SocketAddr],
-        genesis: Block<N>,
-        cdn: Option<String>,
-        dev: Option<u16>,
-    ) -> Result<Self> {
-        Ok(Self::Validator(Arc::new(
-            Validator::new(node_ip, rest_ip, account, trusted_peers, genesis, cdn, dev).await?,
-        )))
-    }
-
-    /// Initializes a new prover node.
-    pub async fn new_prover(
-        node_ip: SocketAddr,
-        account: Account<N>,
-        trusted_peers: &[SocketAddr],
-        genesis: Block<N>,
-        dev: Option<u16>,
-    ) -> Result<Self> {
-        Ok(Self::Prover(Arc::new(Prover::new(node_ip, account, trusted_peers, genesis, dev).await?)))
-    }
-
-    /// Initializes a new client node.
-    pub async fn new_client(
-        node_ip: SocketAddr,
-        account: Account<N>,
-        trusted_peers: &[SocketAddr],
-        genesis: Block<N>,
-        dev: Option<u16>,
-    ) -> Result<Self> {
-        Ok(Self::Client(Arc::new(Client::new(node_ip, account, trusted_peers, genesis, dev).await?)))
-    }
-
-    /// Returns the node type.
-    pub fn node_type(&self) -> NodeType {
-        match self {
-            Self::Beacon(beacon) => beacon.node_type(),
-            Self::Validator(validator) => validator.node_type(),
-            Self::Prover(prover) => prover.node_type(),
-            Self::Client(client) => client.node_type(),
-        }
-    }
-
-    /// Returns the account private key of the node.
-    pub fn private_key(&self) -> &PrivateKey<N> {
-        match self {
-            Self::Beacon(node) => node.private_key(),
-            Self::Validator(node) => node.private_key(),
-            Self::Prover(node) => node.private_key(),
-            Self::Client(node) => node.private_key(),
-        }
-    }
-
-    /// Returns the account view key of the node.
-    pub fn view_key(&self) -> &ViewKey<N> {
-        match self {
-            Self::Beacon(node) => node.view_key(),
-            Self::Validator(node) => node.view_key(),
-            Self::Prover(node) => node.view_key(),
-            Self::Client(node) => node.view_key(),
-        }
-    }
-
-    /// Returns the account address of the node.
-    pub fn address(&self) -> Address<N> {
-        match self {
-            Self::Beacon(node) => node.address(),
-            Self::Validator(node) => node.address(),
-            Self::Prover(node) => node.address(),
-            Self::Client(node) => node.address(),
-        }
-    }
-
-    /// Returns `true` if the node is in development mode.
-    pub fn is_dev(&self) -> bool {
-        match self {
-            Self::Beacon(node) => node.is_dev(),
-            Self::Validator(node) => node.is_dev(),
-            Self::Prover(node) => node.is_dev(),
-            Self::Client(node) => node.is_dev(),
-        }
-    }
-}
